UNTESTED: fexcore: Add guest context reconstruction support
This should allow for guest applications to handle e.g. access violation
exceptions somewhat correctly and observe register state at the time of the
exception. This doesn't correctly handle cases where the guest writes to
registers in its exception handler since it operates by stashing the ARM
context associated with the X86 context on to the stack (as FEX lacks
support for restoring from exceptions that occured partway through an instr rn)