loader: Use RWX rather than RW while processing relocations
The sections of a program are temporarily reprotected to RW while processing
relocations, this is problematic as SELinux doesn't allow any dirty pages to be
reprotected as executable if they weren't executable when the writes took place
under the execmod policy. To fix this, the sections are now reprotected to RWX
instead so the policy isn't triggered during the reprotection to executable.